Completely refactored/rebuilt CLI now working with new app structure

+++ components/studio/api/views.py
@@ -1,7 +1,8 @@
 import uuid
+import json
 from ast import literal_eval
 from django_filters.rest_framework import DjangoFilterBackend
-from django.http import HttpResponse
+from django.http import HttpResponse, JsonResponse
 from django.db.models import Q
 from rest_framework.mixins import CreateModelMixin, ListModelMixin, RetrieveModelMixin, UpdateModelMixin
 from rest_framework.viewsets import GenericViewSet
@@ -11,14 +12,20 @@ from rest_framework import generics
 from .APIpermissions import ProjectPermission
 from deployments.helpers import build_definition
 from projects.helpers import create_project_resources
+from projects.tasks import create_resources_from_template, delete_project_apps
 from django.contrib.auth.models import User
 from django.conf import settings
 import modules.keycloak_lib as kc
-from projects.models import Environment
+from projects.models import Environment, Flavor, S3, MLFlow, ProjectTemplate, ProjectLog
 from models.models import ObjectType
+from apps.models import AppInstance
 
 from .serializers import Model, MLModelSerializer, ModelLog, ModelLogSerializer, Metadata, MetadataSerializer, Project, ProjectSerializer, UserSerializer
-from .serializers import ObjectTypeSerializer
+from .serializers import ObjectTypeSerializer, AppInstanceSerializer, FlavorsSerializer
+from .serializers import EnvironmentSerializer, S3serializer, MLflowSerializer
+
+from projects.tasks import create_resources_from_template
+from apps.tasks import delete_resource
 
 class ObjectTypeList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
     permission_classes = (IsAuthenticated, ProjectPermission,)
@@ -27,10 +34,6 @@ class ObjectTypeList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, Updat
     filterset_fields = ['id','name', 'slug']
 
     def get_queryset(self):
-        """
-        This view should return a list of all the models
-        for the currently authenticated user.
-        """
         return ObjectType.objects.all()
 
 class ModelList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
@@ -47,9 +50,13 @@ class ModelList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateMode
         return Model.objects.filter(project__pk=self.kwargs['project_pk'])
 
     def destroy(self, request, *args, **kwargs):
+        project = Project.objects.get(id=self.kwargs['project_pk'])
         model = self.get_object()
-        model.delete()
-        return HttpResponse('ok', 200)
+        if model.project==project:
+            model.delete()
+            return HttpResponse('ok', 200)
+        else:
+            return HttpResponse('User is not allowed to delete object.', 403)
 
     def create(self, request, *args, **kwargs):
         project = Project.objects.get(id=self.kwargs['project_pk'])
@@ -211,8 +218,37 @@ class ProjectList(generics.ListAPIView, GenericViewSet, CreateModelMixin, Retrie
         current_user = self.request.user
         return Project.objects.filter(Q(owner__username=current_user) | Q(authorized__pk__exact=current_user.pk), ~Q(status='archived'))
     
-    @action(detail=False, methods=['post'], permission_classes=[IsAuthenticated])
-    def create_project(self, request):
+    def destroy(self, request, *args, **kwargs):
+        project = self.get_object()
+        if request.user == project.owner and project.status.lower() != "deleted":
+            print("Delete project")
+
+            print("Cleaning up Keycloak.")
+            keyc = kc.keycloak_init()
+            kc.keycloak_delete_client(keyc, project.slug)
+            
+            scope_id, res = kc.keycloak_get_client_scope_id(keyc, project.slug+'-scope')
+            print(scope_id)
+            kc.keycloak_delete_client_scope(keyc, scope_id)
+
+            print("KEYCLOAK RESOURCES DELETED SUCCESFULLY!")
+            print("SCHEDULING DELETION OF ALL INSTALLED APPS")
+            delete_project_apps(project.slug)
+
+            print("ARCHIVING PROJECT Object")
+            objects = Model.objects.filter(project=project)
+            for obj in objects:
+                obj.status = 'AR'
+                obj.save()
+            project.status = 'archived'
+            project.save()
+        else:
+            print("User is not allowed to delete project (only owner can delete).")
+            return HttpResponse("User is not allowed to delete project (only owner can delete).", status=403)
+
+        return HttpResponse('ok', status=200)
+
+    def create(self, request):
         name = request.data['name']
         description = request.data['description']
         repository = request.data['repository']
@@ -221,13 +257,164 @@ class ProjectList(generics.ListAPIView, GenericViewSet, CreateModelMixin, Retrie
                                                  description=description,
                                                  repository=repository)
         success = True
+        
         try:
-            create_project_resources(project, request.user, repository=repository)
-        except:
+            # Create project resources (Keycloak only)
+            create_project_resources(project, request.user.username, repository)
+
+            # Create resources from the chosen template
+            template_slug = request.data['template']
+            template = ProjectTemplate.objects.get(slug=template_slug)
+            project_template = ProjectTemplate.objects.get(pk=template.pk)
+            create_resources_from_template.delay(request.user.username, project.slug, project_template.template)
+
+            # Reset user token
+            if 'oidc_id_token_expiration' in request.session:
+                request.session['oidc_id_token_expiration'] = time.time()-100
+                request.session.save()
+            else:
+                print("No token to reset.")
+        except Exception as e:
             print("ERROR: could not create project resources")
+            print(e)
             success = False
-            return HttpResponse('Ok', status=400)
+
+        if not success:
+            project.delete()
+            return HttpResponse('Failed to create project.', status=200)
+        else:
+            l1 = ProjectLog(project=project, module='PR', headline='Project created',
+                            description='Created project {}'.format(project.name))
+            l1.save()
+
+            l2 = ProjectLog(project=project, module='PR', headline='Getting started',
+                            description='Getting started with project {}'.format(project.name))
+            l2.save()
+
 
         if success:
             project.save()
-            return HttpResponse('Ok', status=200)
+            return HttpResponse(project.slug, status=200)
+
+
+class ResourceList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
+    permission_classes = (IsAuthenticated, ProjectPermission,)
+    serializer_class = AppInstanceSerializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = ['id','name', 'app__category']
+
+    # def get_queryset(self):
+    #     return AppInstance.objects.filter(~Q(state="Deleted"), project__pk=self.kwargs['project_pk'])
+    
+    def create(self, request, *args, **kwargs):
+        template = request.data
+        # template = {
+        #     "apps": request.data
+        # }
+        print(template)
+        project = Project.objects.get(id=self.kwargs['project_pk'])
+        create_resources_from_template.delay(request.user.username, project.slug, json.dumps(template))
+        return HttpResponse("Submitted request to create app.", status=200)
+
+class AppInstanceList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
+    permission_classes = (IsAuthenticated, ProjectPermission,)
+    serializer_class = AppInstanceSerializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = ['id','name', 'app__category']
+
+    def get_queryset(self):
+        return AppInstance.objects.filter(~Q(state="Deleted"), project__pk=self.kwargs['project_pk'])
+    
+    def create(self, request, *args, **kwargs):
+        return HttpResponse("Use 'resources' endpoint instead.", status=200)
+
+    def destroy(self, request, *args, **kwargs):
+        project = Project.objects.get(id=self.kwargs['project_pk'])
+        appinstance = self.get_object()
+        # Check that user is allowed to delete app:
+        # Either user owns the app, or is a member of the project (Checked by project permission above)
+        # and the app is set to project level permission.
+        access = False
+        if appinstance.owner == request.user:
+            print("User owns app, can delete.")
+            access = True
+        elif appinstance.permission.projects.filter(slug=project.slug).exists():
+            print("Project has permission")
+            access = True
+        elif appinstance.permission.public:
+            print("App is public and user has project permission, allowed to delete.")
+            access = True
+        # Q(owner=request.user) | Q(permission__projects__slug=project.slug) |  Q(permission__public=True)
+        if access:
+            delete_resource.delay(appinstance.pk)
+        else:
+            return HttpResponse("User is not allowed to delete resource.", status=403)
+        return HttpResponse("Deleted app.", status=200)
+
+class FlavorsList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
+    permission_classes = (IsAuthenticated, ProjectPermission,)
+    serializer_class = FlavorsSerializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = ['id','name']
+
+    def get_queryset(self):
+        return Flavor.objects.filter(project__pk=self.kwargs['project_pk'])
+
+    def destroy(self, request, *args, **kwargs):
+        try:
+            obj = self.get_object()
+        except:
+            return HttpResponse("No such object.", status=400)
+        obj.delete()
+        return HttpResponse("Deleted object.", status=200)
+
+class EnvironmentList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
+    permission_classes = (IsAuthenticated, ProjectPermission,)
+    serializer_class = EnvironmentSerializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = ['id','name']
+
+    def get_queryset(self):
+        return Environment.objects.filter(project__pk=self.kwargs['project_pk'])
+    
+    def destroy(self, request, *args, **kwargs):
+        try:
+            obj = self.get_object()
+        except:
+            return HttpResponse("No such object.", status=400)
+        obj.delete()
+        return HttpResponse("Deleted object.", status=200)
+
+class S3List(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
+    permission_classes = (IsAuthenticated, ProjectPermission,)
+    serializer_class = S3serializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = ['id','name', 'host', 'region']
+
+    def get_queryset(self):
+        return S3.objects.filter(project__pk=self.kwargs['project_pk'])
+
+    def destroy(self, request, *args, **kwargs):
+        try:
+            obj = self.get_object()
+        except:
+            return HttpResponse("No such object.", status=400)
+        obj.delete()
+        return HttpResponse("Deleted object.", status=200)
+
+class MLflowList(GenericViewSet, CreateModelMixin, RetrieveModelMixin, UpdateModelMixin, ListModelMixin):
+    permission_classes = (IsAuthenticated, ProjectPermission,)
+    serializer_class = MLflowSerializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = ['id','name']
+
+    def get_queryset(self):
+        return MLFlow.objects.filter(project__pk=self.kwargs['project_pk'])
+
+    def destroy(self, request, *args, **kwargs):
+        try:
+            obj = self.get_object()
+        except:
+            return HttpResponse("No such object.", status=400)
+        obj.delete()
+        return HttpResponse("Deleted object.", status=200)
